VBA styles: Interior, Font and NumberFormat onto CellStyle
Section 1 of issue #58. `Range.Interior`, `Range.Font` and
`Range.NumberFormat`, plus the `RGB` builtin.

The colour conversion is the part #58 called "the single most likely thing to
ship backwards", so it is checked from both sides. VBA's `Color` is a **BGR**
`Long` -- `&HFF0000` is blue -- while `CellStyle` stores `"#RRGGBB"`, and an
implementation that gets this wrong produces a file that opens without
complaint and is the wrong colour. Asking the object model cannot catch that:
both engines would round-trip the same wrong `Long`. So the new
`fuzz/vba_style_probe.py --paint` has Excel *save* a workbook after setting
colours and reads the real ARGB back with openpyxl, which is a channel
neither implementation can talk its way past. 6/6 agree.

Two things the measurements corrected rather than confirmed:

* `ColorIndex` is a **nearest-colour match**, not a lookup. The first
  implementation returned `xlNone` for an off-palette fill; Excel returns the
  nearest slot (`RGB(250,10,10)` -> 3, `RGB(10,200,10)` -> 4). The 56-slot
  palette in `vba/color.rs` was read out of Excel one slot at a time by
  `--palette` rather than transcribed.
* Over a range whose cells disagree, every style property reads `Null`
  **except** `Interior.Color`, which reads `0`. Excel's own asymmetry;
  `style_fold` takes the mixed-value as a parameter rather than inferring it,
  since 0 is also a legitimate uniform answer.

`CellStyle::font_size` widens from `u16` to `f64`: Excel reports `Font.Size`
as a `Double` and `10.5` round-trips, so an integer field would have stored a
silently wrong size. This is a breaking change to a public field, which
pre-1.0 `visi-core` can still take.

`Interior` and `Font` ride on the owning range's handle rather than carrying
coordinates, so a styled range tracks a structural edit for free -- there is a
test that would catch it if they ever stopped.

`NumberFormat` writes `CellStyle::num_format`, the same field date handling
reads. Measured and tested: setting `General` on a date cell leaves the serial
alone and changes only the rendering, which is exactly `core::date`'s model.

25 new tests, each asserting what Excel returned for the same expression.

+++ visi-core/src/core/style.rs
@@ -3,7 +3,9 @@
 use serde::{Deserialize, Serialize};
 
 /// Cell formatting style attributes (font color, background color, font styles, font family, font size).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
+// No `Eq`: `font_size` is an `f64`, matching Excel's `Double`-typed
+// `Font.Size`. `PartialEq` is what the codebase actually uses.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
 pub struct CellStyle {
     /// Font color as Hex (e.g. "#FF0000" or "FF0000") or standard color name ("red", "blue", etc.)
     pub font_color: Option<String>,
@@ -17,8 +19,12 @@ pub struct CellStyle {
     pub underline: Option<bool>,
     /// Font family name (e.g. "Arial", "Calibri", "Courier New")
     pub font_family: Option<String>,
-    /// Font size in points (e.g. 11, 12, 14)
-    pub font_size: Option<u16>,
+    /// Font size in points (e.g. 11, 12, 14.5).
+    ///
+    /// `f64` rather than an integer because Excel's is: `Font.Size` reports
+    /// as a `Double` and a half-point size round-trips (`.Font.Size = 10.5`
+    /// reads back as `10.5`), both measured with `fuzz/vba_style_probe.py`.
+    pub font_size: Option<f64>,
     /// Excel number-format code (e.g. `m/d/yy`, `yyyy-mm-dd`).
     ///
     /// This is how a date cell remembers the notation it was written in: the
@@ -104,7 +110,7 @@ mod tests {
         let style2 = CellStyle {
             bg_color: Some("#00FF00".to_string()),
             font_color: Some("#0000FF".to_string()),
-            font_size: Some(14),
+            font_size: Some(14.0),
             ..Default::default()
         };
 
@@ -112,6 +118,6 @@ mod tests {
         assert_eq!(style1.font_color, Some("#0000FF".to_string()));
         assert_eq!(style1.bg_color, Some("#00FF00".to_string()));
         assert_eq!(style1.bold, Some(true));
-        assert_eq!(style1.font_size, Some(14));
+        assert_eq!(style1.font_size, Some(14.0));
     }
 }
